==> code/05-MergeDeathsEmissions.ipynb <==
# Merge Deaths-Emissions

import pandas as pd
import numpy  as np

DfDeaths = pd.read_csv('../data/Deaths/DfDefuncionesGeo.csv', sep=';', index_col=0)

FEmissions = pd.ExcelFile('../data/csv/Emissions/geo/Comp_by_Emi_total_Geo_2001-2015.xlsx', 
                       sheetname='Comp_by_Emi_total_Geo_2001-2015')
DfEmissionsTotal = pd.read_excel(FEmissions,index_col=None,na_values=['NA'])


### Elimina registro de defunciones por causas no, posiblemente, relacionadas con emisiones
# Según manual CIE-10-ES de diagnósticos de enfermedades del Ministerio de Sanida
# Se Elimina:
# 1.- Enfermedades infecciosas y parasitarias A00-B99
# 2.- Malformaciones congénitas, deformaciones y anomalías cromosomáticas Q00-Q99
# 3.- Lesiones y envenenamientos S00-S99
# 4.- Factores Sanitarios Z00-Z99 
DfDeaths = DfDeaths[(DfDeaths['CausaMuertebas1'] != 'A') & \
                    (DfDeaths['CausaMuertebas1'] != 'Q') & \
                    (DfDeaths['CausaMuertebas1'] != 'S') & \
                    (DfDeaths['CausaMuertebas1'] != 'Z') 
                   ]

### cálculo área posible repercusión de las emisiones 5km
DfDeaths['Longitud1'] = DfDeaths['Longitud'].apply(lambda x:  x + 0.070000000000000)
DfDeaths['Longitud2'] = DfDeaths['Longitud'].apply(lambda x:  x - 0.070000000000000)
DfDeaths['Latitud1'] = DfDeaths['Latitud'].apply(lambda x:  x + 0.070000000000000)
DfDeaths['Latitud2'] = DfDeaths['Latitud'].apply(lambda x:  x - 0.070000000000000)

# Nueva columna Id fichero defunciones para identificar cada defunción
DfDeaths['Id'] = np.arange(len(DfDeaths))

#Selección de Emisiones dentro del área aproximada de 5km cuadrados de defunciones y entre el año de 
#nacimiento y defunción.
#Sólo se tratan las emisiones por aíre.
#Optimización de proceso: tratamiento por años de emisión, liberando memoría y grabando en disco 
#cada 500 registros.
DfDeathsEmissions = pd.DataFrame()
DfEmissions = pd.DataFrame()

for anio in range(2001,2016):
    
    del DfEmissions
    
    DfEmissions = DfEmissionsTotal[(DfEmissionsTotal['MedioReceptor'] == 'Aire') & 
                         (DfEmissionsTotal['AnyoReferencia'] == anio)]
    DfEmissions = DfEmissions.reset_index()
    
    for i in range(1,len(DfEmissions)):
        Df1 = DfDeaths.loc[(DfDeaths['Longitud1'] > DfEmissions.Longitud.at[i]) &
                       (DfDeaths['Longitud2'] < DfEmissions.Longitud.at[i]) &                  
                       (DfDeaths['Latitud1'] > DfEmissions.Latitud.at[i]) &                   
                       (DfDeaths['Latitud2'] < DfEmissions.Latitud.at[i]) &      
                       (DfDeaths['AnioNacimiento'] <= DfEmissions.AnyoReferencia.at[i]) &
                       (DfDeaths['AnioDefuncion']  >= DfEmissions.AnyoReferencia.at[i])]

   #     print(i)
                  
        if len(Df1) > 0:
            Df1['CodigoPRTR'] = DfEmissions.CodigoPRTR.at[i]
            Df1['AnyoReferencia'] = DfEmissions.AnyoReferencia.at[i]
            Df1['Contaminante'] = DfEmissions.Contaminante.at[i]
            Df1['CantidadTotalkgporanyo'] = DfEmissions.CantidadTotalkgporaAnyo.at[i]
            Df1['LongitudE'] = DfEmissions.Longitud.at[i]
            Df1['LatitudE'] = DfEmissions.Latitud.at[i]

            DfDeathsEmissions = DfDeathsEmissions.append(Df1)
    
        if (i == 500) | (i == 1000) | (i == 1500) | (i == 2000) | (i == 2500) | \
           (i == 3000) | \
           (i == 3500):
             DfDeathsEmissions.to_csv('../data/csv/merge/DeathsEmissionsFinal_'+str(anio)+'_'+str(i)+'.csv',
                                      sep=';',compression='gzip')
             del DfDeathsEmissions
             DfDeathsEmissions = pd.DataFrame()
        

    DfDeathsEmissions.to_csv('../data/csv/merge/DeathsEmissionsFinal_'+str(anio)+'_'+str(i)+'.csv',
                             sep=';',compression='gzip')
    del DfDeathsEmissions
    DfDeathsEmissions = pd.DataFrame()
