==> code/14-Model_Spatial_Regression.ipynb <==
# Model Spatial Regression

import pandas as pd
import numpy as np
import csv

#### Librerias específicas para la Autocorrelación espacial con Pysal

#For Spatial Correlation
# !pip install pysal
#http://pysal.readthedocs.io/en/latest/
import pysal as ps
from pysal.spreg import ols
from pysal.spreg import ml_error
from pysal.spreg import ml_lag

Recogemos el CSV procedente del Modelo Knn-vecinos con la variable Contaminante categórica codificada

csv_path_geom = '../data/csv/model_csv/df_model_geom_encoding.csv' #Recogemos el df Con geometrias para leer con geopandas
shp_path = '../data/shapes/Deaths/Deaths2015.shp'

#gdf = gpd.read_file(shp_path)
df = pd.read_csv(csv_path_geom, sep=';', encoding = 'utf-8', compression='gzip', index_col=False)

df.drop(['Unnamed: 0'], axis=1, inplace=True) #Borramos columnas innecesarias

pd.options.display.max_columns = None

df.columns

Observamos que todas las variables son numéricas pero habrá que escalar estas

df.dtypes

Por conveniencia

df.rename(columns={'Cont_Dióxido de carbono (CO2)':'CO2_event'}, inplace=True)
df.rename(columns={'Cont_Partículas (PM10)':'PM10_event'}, inplace=True)

#### Pre-processing: Feature Scaling 

from sklearn.preprocessing import StandardScaler

#Standarization de columnas importantes
columnas = ['AnioCumplidos', 'TamanioMuniResi', 'total_anios_Expo_Id', 'Total_Kg_expo','Habitantes']
scaler = StandardScaler()
df[columnas] = scaler.fit_transform(df[columnas]) #Solo columnas de interés escaladas

df.head()

### Matrices de pesos espaciales (Spatial Weights) 

Debido a que el shape es muy pesado para realizar un análisis de regresión espacial primero vamos a realizar un muestreo aleatorio con numpy

rows = np.random.choice(df.index.values, 5000)

df_sam = df.loc[rows]

df_sam.shape

df_sam.head()

### k-nearest neighbor weights

Este método es una combinación de umbrales basados en la distancia junto con los pesos del kernel. En nuestro caso el ancho de banda se establece en un valor predeterminado y se fija a través de las observaciones

del df_sam['geometry_death']

En principio establecemos una matriz de pesos espaciales para k vecinos

wknn = ps.weights.KNN(df_sam, k = 30)

wknn.s0

wknn.n

Estandarizamos rows

wknn.transform = 'r'

## Modelos de Regresión

df_sam.columns

Generamos los array

columns=['Sexo', 'AnioCumplidos','total_anios_Expo_Id', 'Total_Kg_expo','Habitantes','Distance']#se utiliza más adelante
np.random.seed(12345)
y = np.array(df_sam['target'])
y.shape = (len(y),1)
X= []
X.append(df_sam['Sexo'])
X.append(df_sam['AnioCumplidos'])
X.append(df_sam['total_anios_Expo_Id'])
X.append(df_sam['Total_Kg_expo'])
X.append(df_sam['Habitantes'])
X.append(df_sam['Distance'])

X = np.array(X).T

### Lineal Regression MCO (Nonspatial)

ls = ols.OLS(y, X, name_y = 'Causa fallecimiento CI10-34', name_x = columns\
             , name_ds = 'Dataframe Deaths sampled')
print(ls.summary)

#### Para ver la no independencia espacial, primero utilizamos la función I de Moran con los residuos desde un punto de vista global

Evaluamos la autocorrelación espacial en los residuos (ya que se supone que son independientes y no están correlacionados). 
Los valores de I pueden oscilar entre -1 (indicando dispersión perfecta) a 1 (correlación perfecta). Un valor de cero indica un patrón espacial aleatorio.

# Pasamos los weights matrix a la función pysal.Moran, a través de nuestro modelo residual (ls.u)
mi = ps.Moran(ls.u, wknn, two_tailed=False)
print('Observed I:', mi.I, '\nExpected I:', mi.EI, '\n   p-value:', mi.p_norm)

### Spatial Regressión

Utilizamos la matriz de pesos espaciales k-nearest neighbor weights y la función de spatial Error model para cuantificar la no-independencia espacial

spat_err = ml_error.ML_Error(y, X, wknn, 
                             name_y='Causa fallecimiento CI10-34', name_x= ['Sexo','Años','Total años Expo','Kgs Cont.expo','Distance','Tamaño Municipio'], 
                             name_w='deaths wknn', name_ds='Dataframe Deaths sampled')
print(spat_err.summary)

#### Regresión lineal por Pysal con matriz de pesos espaciales

Realizamos otra Regresión mediante otro método del módulo spreg de Pysal

m1 = ps.spreg.OLS(y, X ,w=wknn, spat_diag=True, name_x=columns, name_y='target',\
                  name_w='deaths wknn',name_ds='Dataframe Deaths sampled')
print(m1.summary)

#### Spatially lagged exogenous regressors

m3 = ps.spreg.GM_Lag(y, X, \
                  w=wknn, spat_diag=True, \
                  name_x=columns, name_y = 'Causa fallecimiento CI10-34',\
                     name_w='deaths wknn',name_ds='Dataframe Deaths sampled')
print(m3.summary)
